@@ -29,11 +29,11 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@babel/core": "^7.3.3",
-    "@babel/plugin-proposal-class-properties": "^7.3.3",
+    "@babel/core": "^7.3.4",
+    "@babel/plugin-proposal-class-properties": "^7.3.4",
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@babel/plugin-transform-runtime": "^7.2.0",
-    "@babel/preset-env": "^7.3.1",
+    "@babel/plugin-transform-runtime": "^7.3.4",
+    "@babel/preset-env": "^7.3.4",
     "@babel/preset-react": "^7.0.0",
     "@babel/register": "^7.0.0",
     "babel-eslint": "^10.0.1",
@@ -43,7 +43,7 @@
     "css-loader": "^2.1.0",
     "cypress": "^3.1.5",
     "cypress-testing-library": "^2.3.6",
-    "eslint": "^5.14.1",
+    "eslint": "^5.15.1",
     "eslint-config-airbnb": "^17.1.0",
     "eslint-config-with-prettier": "^6.0.0",
     "eslint-loader": "^2.1.2",
@@ -63,24 +63,24 @@
     "npm-merge-driver": "^2.3.5",
     "optimize-css-assets-webpack-plugin": "^5.0.1",
     "postcss": "^7.0.14",
-    "postcss-apply": "^0.11.0",
+    "postcss-apply": "^0.12.0",
     "postcss-import": "^12.0.1",
     "postcss-loader": "^3.0.0",
-    "postcss-nested": "^4.1.1",
-    "postcss-preset-env": "^6.5.0",
+    "postcss-nested": "^4.1.2",
+    "postcss-preset-env": "^6.6.0",
     "raw-loader": "^1.0.0",
-    "react-test-renderer": "^16.8.2",
+    "react-test-renderer": "^16.8.3",
     "start-server-and-test": "^1.7.11",
     "style-loader": "^0.23.1",
-    "uglifyjs-webpack-plugin": "^2.1.1",
+    "uglifyjs-webpack-plugin": "^2.1.2",
     "url-loader": "^1.1.2",
-    "webpack": "^4.29.5",
+    "webpack": "^4.29.6",
     "webpack-cli": "^3.2.3",
-    "webpack-dev-server": "^3.2.0"
+    "webpack-dev-server": "^3.2.1"
   },
   "dependencies": {
     "@babel/polyfill": "^7.2.5",
-    "@babel/runtime": "^7.3.1",
+    "@babel/runtime": "^7.3.4",
     "axios": "^0.18.0",
     "axios-mock-adapter": "^1.16.0",
     "babel-plugin-module-resolver": "^3.2.0",
@@ -88,8 +88,8 @@
     "clsx": "^1.0.3",
     "copy-webpack-plugin": "^5.0.0",
     "enzyme": "^3.9.0",
-    "enzyme-adapter-react-16": "^1.9.1",
-    "eslint-config-prettier": "^4.0.0",
+    "enzyme-adapter-react-16": "^1.10.0",
+    "eslint-config-prettier": "^4.1.0",
     "eslint-import-resolver-webpack": "^0.11.0",
     "jest": "^24.1.0",
     "jest-dom": "^3.1.2",
@@ -98,14 +98,14 @@
     "path-to-regexp": "^3.0.0",
     "prop-types": "^15.7.2",
     "qs": "^6.6.0",
-    "react": "^16.8.2",
-    "react-dom": "^16.8.2",
-    "react-dropzone": "^8.2.0",
+    "react": "^16.8.3",
+    "react-dom": "^16.8.3",
+    "react-dropzone": "^9.0.0",
     "react-google-maps": "^9.4.5",
     "react-redux": "^6.0.1",
     "react-router-dom": "4.4.0-beta.6",
     "react-select": "^2.4.1",
-    "react-testing-library": "^5.9.0",
+    "react-testing-library": "^6.0.0",
     "react-textarea-autosize": "^7.1.0",
     "redux": "^4.0.1",
     "redux-thunk": "^2.3.0"
