@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useEffect } from 'react'
 import { Route, withRouter } from 'react-router-dom'
 import { connect } from 'react-redux'
 import { object, func } from 'prop-types'
@@ -8,7 +8,7 @@ import MainMenu from '@sdog/scenes/app/menu'
 import DashFooter from '@sdog/scenes/app/footer'
 import UserMenu from '@sdog/components/user_menu'
 import Spinner from '@sdog/components/spinner'
-import { getUserProfile } from '@sdog/store/user'
+import { getUserProfile as getUserProfileAction } from '@sdog/store/user'
 
 import theme from './theme.css'
 import './styles.css'
@@ -17,44 +17,40 @@ const DashboardScene = React.lazy(() => import('@sdog/scenes/dashboard'))
 const SearchScene = React.lazy(() => import('@sdog/scenes/search'))
 const SettingsScene = React.lazy(() => import('@sdog/scenes/settings'))
 
-class App extends React.Component {
-  componentDidMount() {
+const App = ({ getUserProfile, location }) => {
+  useEffect(() => {
     setHtmlClass('html-app')
-    this.props.getUserProfile()
-  }
+    getUserProfile()
 
-  componentWillUnmount() {
-    removeHtmlClass('html-app')
-  }
+    return removeHtmlClass('html-app')
+  }, [])
 
-  render() {
-    return (
-      <div className={theme.app}>
-        <header className={theme.appHeader}>
-          <div className={theme.appHeaderInner}>
-            <div className={theme.logo}>
-              <Logo hideTextOnMobile />
-            </div>
-
-            <MainMenu location={this.props.location} />
-
-            <div className={theme.userMenu}>
-              <UserMenu />
-            </div>
+  return (
+    <div className={theme.app}>
+      <header className={theme.appHeader}>
+        <div className={theme.appHeaderInner}>
+          <div className={theme.logo}>
+            <Logo hideTextOnMobile />
           </div>
-        </header>
 
-        <div className={theme.appContent}>
-          <React.Suspense fallback={<Spinner />}>
-            <Route path="/" component={DashboardScene} exact />
-            <Route path="/settings" component={SettingsScene} />
-            <Route path="/search" component={SearchScene} />
-          </React.Suspense>
+          <MainMenu location={location} />
+
+          <div className={theme.userMenu}>
+            <UserMenu />
+          </div>
         </div>
-        <DashFooter />
+      </header>
+
+      <div className={theme.appContent}>
+        <React.Suspense fallback={<Spinner />}>
+          <Route path="/" component={DashboardScene} exact />
+          <Route path="/settings" component={SettingsScene} />
+          <Route path="/search" component={SearchScene} />
+        </React.Suspense>
       </div>
-    )
-  }
+      <DashFooter />
+    </div>
+  )
 }
 
 App.propTypes = {
@@ -62,7 +58,7 @@ App.propTypes = {
   getUserProfile: func.isRequired,
 }
 
-export const mapActionsToProps = { getUserProfile }
+export const mapActionsToProps = { getUserProfile: getUserProfileAction }
 
 export default withRouter(
   connect(
